fix site build by disabling pre-render of /mp-[slug] page
- Rename `mp_build_bucket` to `mp_bucket` for consistency.
- Introduce `fetch_material_data` function for parallel loading from different mp_buckets
- Add a new event handler `emit_file_load_event` to streamline file load events in `Structure.svelte`.
- Remove deprecated test cases and consolidate pbc/math/structure util tests
